chromium/third_party/blink/web_tests/virtual/out-of-blink-cspee/external/wpt/content-security-policy/embedded-enforcement/subsumption_algorithm-host_sources-protocols-expected.txt

This is a testharness.js-based test.
FAIL `https` is more restrictive than `http`. assert_equals: expected (undefined) undefined but got (boolean) true
PASS The reverse allows iframe be to be loaded.
PASS Matching `https` protocols.
PASS `http:` should subsume all host source expressions with this protocol.
PASS `http:` should subsume all host source expressions with `https:`.
FAIL `http:` does not subsume other protocols. assert_equals: expected (undefined) undefined but got (boolean) true
FAIL If scheme source is present in returned csp, it must be specified in required csp too. assert_equals: expected (undefined) undefined but got (boolean) true
PASS `http:` subsumes other `http:` source expression.
PASS `http:` subsumes other `https:` source expression and expressions with `http:`.
FAIL All scheme sources must be subsumed. assert_unreached: No message should be sent from the frame. Reached unreachable code
PASS All scheme sources are subsumed by their stronger variants.
Harness: the test ran to completion.
